requirements.txt
# This is the main requirements.txt file that includes all python dependencies
# for the entire project. There are also client requirements files that are
# maintained independently to minimise the installs required to build the
# respective clients:
#
# - client/windows/requirements.txt
# - client/linux/requirements.txt (also used for OS X)
#
# Ideally *this* file would just contain server dependencies and use '-r' to
# import client dependencies. This is impossible since the same dependency
# cannot be listed more than once:
# https://github.com/pypa/pip/issues/56
#
# and we have multiple packages that are required in both the server and the
# client. So until pip can handle proper dependency resolution we'll maintain
# these separately. Please check if the client files need to be updated if you
# update this one.
Django==1.8.3
PyYAML==3.11
argparse==1.4.0
binplist==0.1.4
gnureadline==6.3.3
google-api-python-client==1.4.2
ipaddr==2.1.11
ipython==4.1.1
matplotlib==1.5.1
mock==1.3.0
mox==0.5.3
pexpect==4.0.1
portpicker==1.1.1
psutil==3.4.2
pyaml==15.8.2
pycrypto==2.6.1
pyinstaller==3.1.1
python-crontab==2.0.1
python-dateutil==2.4.2
pytz==2015.7
http://images.rekall-forensic.com/share/rekall-core-1.5.0.tar.gz#egg=rekall-core-1.5
selenium==2.50.1
urllib3==1.14
wsgiref==0.1.2
Werkzeug==0.11.3
GRR-M2Crypto==0.22.4
pytsk3==20160207.post1
MySQL-python==1.2.5
winjob==0.0.11
